+++ bin/scene-capability-engine.js
@@ -24,6 +24,7 @@ const { registerSpecRelatedCommand } = require('../lib/commands/spec-related');
 const { registerTimelineCommands } = require('../lib/commands/timeline');
 const { registerValueCommands } = require('../lib/commands/value');
 const { registerTaskCommands } = require('../lib/commands/task');
+const { registerAuthCommands } = require('../lib/commands/auth');
 const VersionChecker = require('../lib/version/version-checker');
 const {
   findLegacyKiroDirectories,
@@ -945,6 +946,7 @@ registerOrchestrateCommands(program);
 // Value realization and observability commands
 registerValueCommands(program);
 registerTaskCommands(program);
+registerAuthCommands(program);
 
 // Template management commands
 const templatesCommand = require('../lib/commands/templates');